Moved nxreduce.Lock into its own file to remove circular dependencies
