+++ setup.py
@@ -56,7 +56,8 @@ setup (name = nxrefine.__package_name__,
                                 'nxcombine=nxrefine.scripts.nxcombine:main',
                                 'nxtransform=nxrefine.scripts.nxtransform:main',
                                 'nxreduce=nxrefine.scripts.nxreduce:main',
-                                'nxsummary=nxrefine.scripts.nxsummary:main'],
+                                'nxsummary=nxrefine.scripts.nxsummary:main',
+                                'nxdatabase=nxrefine.scripts.nxdatabase:main'],
        },
 #       data_files=[('', ['bm/b1.gif', 'bm/b2.gif']),
 #                  ('config', ['cfg/data.cfg']),
